+++ ProjetAgenceVoyage/src/main/java/fr/adaming/model/Service.java
@@ -24,17 +24,17 @@ public class Service implements Serializable {
 		this.questionnaire = questionnaire;
 	}
 
-	public Service(int id, String questionnaire) {
+	public Service(Long id, String questionnaire) {
 		super();
 		this.id = id;
 		this.questionnaire = questionnaire;
 	}
 
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
